Move MRCP helpers into dedicated contrib package
