+++ opengate/contrib/__init__.py
@@ -1,1 +1,1 @@
-from . import dose, linacs, pet, phantoms, spect, tps, beamlines, optical
+from . import dose, linacs, mrcp, pet, phantoms, spect, tps, beamlines, optical
