+++ opengate/contrib/mrcp/__init__.py
@@ -1,0 +1,1 @@
+from . import mrcp, mrcp_utils
