--- opengate/contrib/mrcp/test
@@ -1,1 +1,0 @@
-
